+++ cosmoscx/src/result.rs
@@ -45,8 +45,12 @@ pub enum ResultCode {
     /// See [`ErrorKind::InvalidQuery`].
     InvalidQuery = -11,
 
+    /// See [`ErrorKind::UnsupportedFeature`].
+    UnsupportedFeature = -12,
+
+    
     /// See [`ErrorKind::IllegalArgumentError`].
-    IllegalArgumentError = -12,
+    IllegalArgumentError = -13,
 }
 
 impl From<azure_data_cosmos_engine::Error> for ResultCode {
@@ -71,6 +75,7 @@ impl From<ErrorKind> for ResultCode {
             ErrorKind::InvalidRequestId => ResultCode::InvalidRequestId,
             ErrorKind::InvalidQuery => ResultCode::InvalidQuery,
             ErrorKind::PythonError => ResultCode::InternalError,
+            ErrorKind::UnsupportedFeature => ResultCode::UnsupportedFeature,
             ErrorKind::IllegalArgumentError => ResultCode::IllegalArgumentError,
         }
     }