+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.4.0 (2025-11-24)
 
 ### Features Added
-* Added ability to execute `readMany` requests against the client engine. - [PR 53](https://github.com/Azure/azure-cosmos-client-engine/pull/53)
+* Added ability to execute `readMany` requests against the client engine to read multiple items. Results are not guaranteed to be ordered. - [PR 53](https://github.com/Azure/azure-cosmos-client-engine/pull/53)
 
 ## 0.3.0 (2025-11-20)